+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+- Fix issue #100: modify to catch Textrepo::MissingTimestampError.
 
 ## [0.4.13] - 2021-03-30
 ### Changed